feat: fix floating window border and add proxy settings
- Fix transparent background for floating window by setting initial
  CSS background to transparent before JS mounts
- Fix Windows build error (DWMNCRP_DISABLED i32->u32 cast)
- Add network proxy settings (none/system/custom HTTP/SOCKS5)
- All reqwest clients now respect proxy configuration
- Add proxy UI section in Settings page

bump version to 0.1.33

+++ src-tauri/src/lib.rs
@@ -62,7 +62,7 @@ fn disable_floating_window_border(win: &tauri::WebviewWindow) {
 
     if let Ok(hwnd) = win.hwnd() {
         let border_color: u32 = DWMWA_COLOR_NONE;
-        let no_nc_rendering: u32 = DWMNCRP_DISABLED;
+        let no_nc_rendering: u32 = DWMNCRP_DISABLED as u32;
         unsafe {
             let _ = DwmSetWindowAttribute(
                 hwnd.0 as _,
